requirements.txt
boto3>=1.9.91
python-dotenv
